pundit added to friend model

+++ app/controllers/friends_controller.rb
@@ -3,18 +3,24 @@ class FriendsController < ApplicationController
 
   def index
     @friends = current_user.friends
+    @friends = policy_scope(Friend)
+    @users = User.all.where.not(id: current_user.id)
   end
 
   def create
-    friend_full_name = params[:friend_full_name].strip
+    # friend_full_name = params[:friend_full_name].strip
 
-    friend = User.find_by("LOWER(CONCAT(first_name, ' ', last_name)) = ?", friend_full_name.downcase)
+    # friend = User.find_by("LOWER(CONCAT(first_name, ' ', last_name)) = ?", friend_full_name.downcase)
+    @attendee = User.find(params[:attendee_id])
+    # @friend = Friend.new
 
-    if friend
-      if current_user.friends.exists?(attendee_id: friend.id)
+    authorize(Friend)
+
+    if @attendee
+      if current_user.friends.exists?(attendee_id: @attendee.id)
         redirect_to friends_path, notice: "Friend already added."
       else
-        current_user.friends.create(attendee_id: friend.id)
+        current_user.friends.create(attendee_id: @attendee.id)
         redirect_to friends_path, notice: "Friend added successfully."
       end
     else
@@ -25,6 +31,8 @@ class FriendsController < ApplicationController
 
 
   def destroy
+    authorize(@friend)
+
     friend = current_user.friends.find_by(attendee_id: params[:id])
     if friend
       friend.destroy
